+++ lib/recording_studio_accessible/services/grant_recording_access.rb
@@ -4,6 +4,7 @@ module RecordingStudioAccessible
   module Services
     class GrantRecordingAccess < BaseService
       include AccessRecordLifecycle
+      include AccessGrantWriter
 
       def initialize(recording:, actor:, role:, manager_actor: nil, controller: nil)
         @recording = recording
@@ -45,44 +46,6 @@ module RecordingStudioAccessible
         true
       end
 
-      def upsert_access_recording!
-        RecordingStudio::Recording.transaction do
-          lock_grant_scope!
-
-          existing_recordings = existing_access_recordings.to_a
-          access_recording = existing_recordings.first
-
-          if access_recording
-            update_existing_access_recording(access_recording, existing_recordings)
-          else
-            create_access_recording
-          end
-        end
-      end
-
-      def update_existing_access_recording(access_recording, existing_recordings)
-        deduplicate_access_recordings!(existing_recordings.drop(1))
-
-        RecordingStudioAccessible::AccessCreationContext.allow do
-          root_recording.revise(access_recording, actor: manager_actor) do |access|
-            access.role = @role
-          end
-        end
-      end
-
-      def create_access_recording
-        RecordingStudioAccessible::AccessCreationContext.allow do
-          root_recording.record(
-            RecordingStudio::Access,
-            actor: manager_actor,
-            parent_recording: @recording
-          ) do |access|
-            access.actor = @actor
-            access.role = @role
-          end
-        end
-      end
-
       def service_args
         {
           recording_id: @recording&.id,
@@ -98,26 +61,6 @@ module RecordingStudioAccessible
 
       def allowed_access_actor_type?
         RecordingStudioAccessible.configuration.allowed_access_actor_type?(@actor)
-      end
-
-      def root_recording
-        RecordingStudio.root_recording_or_self(@recording)
-      end
-
-      def lock_grant_scope!
-        @recording.lock!
-      end
-
-      def existing_access_recordings
-        return RecordingStudio::Recording.none unless @actor
-
-        RecordingStudioAccessible::DirectAccessQuery.access_recordings_for_actor(recording: @recording, actor: @actor)
-      end
-
-      def deduplicate_access_recordings!(access_recordings)
-        access_recordings.each do |access_recording|
-          destroy_access_recording!(access_recording, manager_actor: manager_actor)
-        end
       end
     end
   end